// libs/dynamic-dto/src/lib/core/interfaces/schema/primitive/string-field.schema.ts
export enum CaseTransform {
  NONE = 'none',
  LOWER = 'lower',
  UPPER = 'upper',
  TITLE = 'title',
  SENTENCE = 'sentence',
  CAMEL = 'camel',
  PASCAL = 'pascal',
  SNAKE = 'snake',
  KEBAB = 'kebab',
  CONSTANT = 'constant',
  DOT = 'dot',
  PATH = 'path',
  HEADER = 'header',
  ALTERNATING = 'alternating',
  INVERSE = 'inverse',
  CAPITALIZE_FIRST = 'capitalize_first',
  CAPITALIZE_WORDS = 'capitalize_words',
}